refactor: rename the _renderer subpackage to _apiref
Includes the one stale test import the rename missed.
