+++ great_docs/_apiref/_render/api_page.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from functools import cached_property
 from typing import TYPE_CHECKING, cast
 
-from great_docs._renderer._render.mixin_page import RenderPageMixin
+from great_docs._apiref._render.mixin_page import RenderPageMixin
 
 from .._format import markdown_escape
 from ..pandoc.blocks import (
