+++ great_docs/core.py
@@ -6659,7 +6659,7 @@ class GreatDocs:
             try:
                 from functools import partial
 
-                from great_docs._renderer.introspect import get_object as qd_get_object
+                from great_docs._apiref.introspect import get_object as qd_get_object
 
                 gd_get_object = partial(qd_get_object, dynamic=True, parser="numpy")
             except ImportError:  # pragma: no cover
@@ -6936,7 +6936,7 @@ class GreatDocs:
         try:
             import griffe
 
-            from great_docs._renderer.introspect import get_object as qd_get_object
+            from great_docs._apiref.introspect import get_object as qd_get_object
         except ImportError:  # pragma: no cover
             # If renderer isn't available, default to True (will fail at build time anyway)
             return True  # pragma: no cover
@@ -7422,14 +7422,14 @@ class GreatDocs:
             try:
                 from functools import partial
 
-                from great_docs._renderer._griffe import (
+                from great_docs._apiref._griffe import (
                     GriffeLoader,
                     LinesCollection,
                     ModulesCollection,
                     Parser,
                 )
-                from great_docs._renderer.introspect import get_parser_defaults
-                from great_docs._renderer.introspect import (
+                from great_docs._apiref.introspect import get_parser_defaults
+                from great_docs._apiref.introspect import (
                     get_object as qd_get_object,
                 )
 
@@ -7664,7 +7664,7 @@ class GreatDocs:
                                     except Exception:  # pragma: no cover
                                         # Dynamic mode failed; try static (no dynamic=True)
                                         try:
-                                            from great_docs._renderer.introspect import (
+                                            from great_docs._apiref.introspect import (
                                                 get_object as qd_get,
                                             )
 
@@ -7731,7 +7731,7 @@ class GreatDocs:
                                                         except Exception:  # pragma: no cover
                                                             # Dynamic failed; try static
                                                             try:
-                                                                from great_docs._renderer.introspect import (
+                                                                from great_docs._apiref.introspect import (
                                                                     get_object as qd_get,
                                                                 )
 
@@ -15636,7 +15636,7 @@ body-classes: "gd-homepage"
 
                 quarto_yml = self.project_path / "_quarto.yml"
                 try:
-                    from great_docs._renderer.api_reference import APIReference
+                    from great_docs._apiref.api_reference import APIReference
 
                     with _quiet_prints():
                         APIReference(str(quarto_yml)).build()
