+++ tests/test_great_docs.py
@@ -30,10 +30,10 @@ from yaml12 import format_yaml as _format_yaml, parse_yaml as _parse_yaml
 
 from great_docs import Config, create_default_config, GreatDocs, load_config
 from great_docs._directives import DocDirectives, extract_directives
-from great_docs._renderer import _globals
-from great_docs._renderer import content
-from great_docs._renderer import spec
-from great_docs._renderer._ast import (
+from great_docs._apiref import _globals
+from great_docs._apiref import content
+from great_docs._apiref import spec
+from great_docs._apiref._ast import (
     DocstringSectionKindPatched,
     DocstringSectionNotes,
     DocstringSectionSeeAlso,
@@ -45,7 +45,7 @@ from great_docs._renderer._ast import (
     tuple_to_data,
     _DocstringSectionPatched,
 )
-from great_docs._renderer._format import (
+from great_docs._apiref._format import (
     format_name,
     format_see_also,
     format_str,
@@ -54,13 +54,13 @@ from great_docs._renderer._format import (
     HAS_RUFF,
     repr_obj,
 )
-from great_docs._renderer._globals import (
+from great_docs._apiref._globals import (
     EXCLUDE_ATTRIBUTES,
     EXCLUDE_CLASSES,
     EXCLUDE_FUNCTIONS,
     EXCLUDE_PARAMETERS,
 )
-from great_docs._renderer._griffe import (
+from great_docs._apiref._griffe import (
     AliasResolutionError,
     dataclasses as dc,
     dataclasses as gdc,
@@ -68,23 +68,23 @@ from great_docs._renderer._griffe import (
     docstrings as gds,
     expressions as expr,
 )
-from great_docs._renderer._griffe.docstrings import DCDocstringSection
-from great_docs._renderer._render import (
+from great_docs._apiref._griffe.docstrings import DCDocstringSection
+from great_docs._apiref._render import (
     get_render_type,
     RenderDocAttribute,
     RenderDocClass,
     RenderDocFunction,
     RenderDocModule,
 )
-from great_docs._renderer._render._label import (
+from great_docs._apiref._render._label import (
     get_label,
     _attribute_label,
     _class_label,
     _function_label,
 )
-from great_docs._renderer._render.api_page import RenderAPIPage
-from great_docs._renderer._render.base import RenderBase
-from great_docs._renderer._render.extending import (
+from great_docs._apiref._render.api_page import RenderAPIPage
+from great_docs._apiref._render.base import RenderBase
+from great_docs._apiref._render.extending import (
     exclude_attributes,
     exclude_classes,
     exclude_functions,
@@ -92,14 +92,14 @@ from great_docs._renderer._render.extending import (
     extend_base_class,
     set_class_attr,
 )
-from great_docs._renderer._render.mixin_members import (
+from great_docs._apiref._render.mixin_members import (
     RenderedMemberPagesGroup,
     RenderedMembersGroup,
 )
-from great_docs._renderer._render.mixin_page import RenderPageMixin
-from great_docs._renderer._render.reference_page import RenderReferencePage
-from great_docs._renderer._render.reference_section import RenderReferenceSection
-from great_docs._renderer._rst_converters import (
+from great_docs._apiref._render.mixin_page import RenderPageMixin
+from great_docs._apiref._render.reference_page import RenderReferencePage
+from great_docs._apiref._render.reference_section import RenderReferenceSection
+from great_docs._apiref._rst_converters import (
     escape,
     sanitize,
     _convert_bold_section_headers,
@@ -121,9 +121,9 @@ from great_docs._renderer._rst_converters import (
     _rst_simple_table_to_md,
     _smart_dedent,
 )
-from great_docs._renderer._tools import render_code_variable, render_type_object
-from great_docs._renderer._visitor import ctx_node, Node, ObjectNotFoundError
-from great_docs._renderer._type_checks import (
+from great_docs._apiref._tools import render_code_variable, render_type_object
+from great_docs._apiref._visitor import ctx_node, Node, ObjectNotFoundError
+from great_docs._apiref._type_checks import (
     griffe_to_doc,
     is_field_init_false,
     is_initvar,
@@ -133,7 +133,7 @@ from great_docs._renderer._type_checks import (
     isDoc,
     no_init,
 )
-from great_docs._renderer.resolve import (
+from great_docs._apiref.resolve import (
     resolve as _blueprint,
     resolve as blueprint_func,
     _Resolver as BlueprintTransformer,
@@ -144,15 +144,15 @@ from great_docs._renderer.resolve import (
     _resolve_alias,
     _to_simple_dict,
 )
-from great_docs._renderer.collect import (
+from great_docs._apiref.collect import (
     build_manifest,
     _ManifestBuilder as CollectTransformer,
     _PackagePrefixRemover as _PagePackageStripper,
     remove_package_prefix,
 )
-from great_docs._renderer.spec import SpecSection
-from great_docs._renderer.api_reference import APIReference, Settings
-from great_docs._renderer.introspect import (
+from great_docs._apiref.spec import SpecSection
+from great_docs._apiref.api_reference import APIReference, Settings
+from great_docs._apiref.introspect import (
     dynamic_alias,
     get_object,
     get_parser_defaults,
@@ -160,16 +160,16 @@ from great_docs._renderer.introspect import (
     _is_valueless,
     _resolve_target,
 )
-from great_docs._renderer import write
-from great_docs._renderer.write import _insert_contents, merge_frontmatter as _merge_frontmatter
-from great_docs._renderer.inventory import (
+from great_docs._apiref import write
+from great_docs._apiref.write import _insert_contents, merge_frontmatter as _merge_frontmatter
+from great_docs._apiref.inventory import (
     convert_inventory,
     create_inventory,
     _create_inventory_item,
     _maybe_call,
 )
-from great_docs._renderer._walkable import MISSING, _Walkable
-from great_docs._renderer.content import (
+from great_docs._apiref._walkable import MISSING, _Walkable
+from great_docs._apiref.content import (
     Doc,
     DocAttribute,
     DocClass,
@@ -181,11 +181,11 @@ from great_docs._renderer.content import (
     Section,
     SummaryItem,
 )
-from great_docs._renderer.inventory import InventoryItem
-from great_docs._renderer.spec import ChildrenStyle
-from great_docs._renderer.spec import SpecObject as Auto
-from great_docs._renderer.spec import SpecOptions as AutoOptions
-from great_docs._renderer.pandoc.blocks import (
+from great_docs._apiref.inventory import InventoryItem
+from great_docs._apiref.spec import ChildrenStyle
+from great_docs._apiref.spec import SpecObject as Auto
+from great_docs._apiref.spec import SpecOptions as AutoOptions
+from great_docs._apiref.pandoc.blocks import (
     Block,
     blockcontent_to_str,
     blockcontent_to_str_items,
@@ -202,8 +202,8 @@ from great_docs._renderer.pandoc.blocks import (
     RawHTMLBlockTag,
     RenderedDocObject,
 )
-from great_docs._renderer.pandoc.components import Attr
-from great_docs._renderer.pandoc.inlines import (
+from great_docs._apiref.pandoc.components import Attr
+from great_docs._apiref.pandoc.inlines import (
     Code,
     Emph,
     inlinecontent_to_str,
@@ -213,7 +213,7 @@ from great_docs._renderer.pandoc.inlines import (
     Str,
     Strong,
 )
-from great_docs._renderer.typing_information import TypeInformation, TypeSections
+from great_docs._apiref.typing_information import TypeInformation, TypeSections
 from great_docs.cli import (
     _detect_optional_dependencies,
     _detect_package_manager,
@@ -14966,8 +14966,8 @@ def test_build_github_source_url_absolute_path():
 
 def _render_doc_source_relative_path(obj):
     """Invoke the renderer's `_source_relative_path` with a stand-in griffe object."""
-    import great_docs._renderer._render.doc as docmod
-    from great_docs._renderer import _type_checks
+    import great_docs._apiref._render.doc as docmod
+    from great_docs._apiref import _type_checks
 
     _type_checks.package_info.cache_clear()
     cls = vars(docmod)["__RenderDoc"]
@@ -21390,7 +21390,7 @@ def test_detect_dynamic_mode_with_cyclic_error():
 
         with (
             patch("griffe.load", return_value=mock_pkg),
-            patch("great_docs._renderer.introspect.get_object", side_effect=fake_get_object),
+            patch("great_docs._apiref.introspect.get_object", side_effect=fake_get_object),
         ):
             result = docs._detect_dynamic_mode("mypkg")
 
@@ -24023,7 +24023,7 @@ def test_build_prepare_and_render_flow():
             patch.object(docs, "_copy_assets", return_value=False),
             patch.object(docs, "_get_quarto_env", return_value={}),
             patch(
-                "great_docs._renderer.api_reference.APIReference",
+                "great_docs._apiref.api_reference.APIReference",
                 mock_builder_class,
             ),
             patch("subprocess.Popen") as mock_popen,
@@ -24127,7 +24127,7 @@ def test_build_dynamic_fallback_to_static():
             patch.object(docs, "_copy_assets", return_value=False),
             patch.object(docs, "_get_quarto_env", return_value={}),
             patch(
-                "great_docs._renderer.api_reference.APIReference",
+                "great_docs._apiref.api_reference.APIReference",
                 mock_builder_class,
             ),
             patch("subprocess.Popen") as mock_popen,
@@ -24182,7 +24182,7 @@ def test_build_static_mode_failure_exits():
             patch.object(docs, "_copy_assets", return_value=False),
             patch.object(docs, "_get_quarto_env", return_value={}),
             patch(
-                "great_docs._renderer.api_reference.APIReference",
+                "great_docs._apiref.api_reference.APIReference",
                 mock_builder_class,
             ),
         ):
@@ -24227,7 +24227,7 @@ def test_build_non_dynamic_failure_exits():
             patch.object(docs, "_copy_assets", return_value=False),
             patch.object(docs, "_get_quarto_env", return_value={}),
             patch(
-                "great_docs._renderer.api_reference.APIReference",
+                "great_docs._apiref.api_reference.APIReference",
                 mock_builder_class,
             ),
         ):
@@ -28363,13 +28363,13 @@ def test_get_object_deprecated_object_name_warns():
 
 def test_get_object_with_shared_loader():
     """get_object reuses a loader if provided."""
-    from great_docs._renderer._griffe import (
+    from great_docs._apiref._griffe import (
         GriffeLoader,
         LinesCollection,
         ModulesCollection,
         Parser,
     )
-    from great_docs._renderer.introspect import get_object
+    from great_docs._apiref.introspect import get_object
 
     loader = GriffeLoader(
         docstring_parser=Parser("numpy"),
@@ -28541,7 +28541,7 @@ def test_replace_docstring_alias():
 
 def test_canonical_path_module():
     """_canonical_path for a module returns module name."""
-    from great_docs._renderer.introspect import _canonical_path
+    from great_docs._apiref.introspect import _canonical_path
 
     result = _canonical_path(json, "")
 
@@ -28550,7 +28550,7 @@ def test_canonical_path_module():
 
 def test_canonical_path_module_with_qualname():
     """_canonical_path for a module with qualname appends suffix."""
-    from great_docs._renderer.introspect import _canonical_path
+    from great_docs._apiref.introspect import _canonical_path
 
     result = _canonical_path(json, "dumps")
 
@@ -28559,7 +28559,7 @@ def test_canonical_path_module_with_qualname():
 
 def test_canonical_path_function():
     """_canonical_path for a function returns module:qualname."""
-    from great_docs._renderer.introspect import _canonical_path
+    from great_docs._apiref.introspect import _canonical_path
 
     result = _canonical_path(json.dumps, "")
     assert result == "json:dumps"
@@ -28567,7 +28567,7 @@ def test_canonical_path_function():
 
 def test_canonical_path_function_with_qualname():
     """_canonical_path for a function with extra qualname appends it."""
-    from great_docs._renderer.introspect import _canonical_path
+    from great_docs._apiref.introspect import _canonical_path
 
     result = _canonical_path(json.dumps, "extra")
     assert result == "json:dumps.extra"
@@ -28575,7 +28575,7 @@ def test_canonical_path_function_with_qualname():
 
 def test_canonical_path_class():
     """_canonical_path for a class returns module:qualname."""
-    from great_docs._renderer.introspect import _canonical_path
+    from great_docs._apiref.introspect import _canonical_path
 
     result = _canonical_path(json.JSONEncoder, "")
     assert result == "json.encoder:JSONEncoder"
@@ -28583,7 +28583,7 @@ def test_canonical_path_class():
 
 def test_canonical_path_plain_object():
     """_canonical_path for a plain object (not module/class/function) returns None."""
-    from great_docs._renderer.introspect import _canonical_path
+    from great_docs._apiref.introspect import _canonical_path
 
     result = _canonical_path(42, "")
     assert result is None
@@ -28591,7 +28591,7 @@ def test_canonical_path_plain_object():
 
 def test_canonical_path_no_module_attr():
     """_canonical_path returns None when __module__ is missing."""
-    from great_docs._renderer.introspect import _canonical_path
+    from great_docs._apiref.introspect import _canonical_path
 
     class NoModule:
         pass
@@ -28604,7 +28604,7 @@ def test_canonical_path_no_module_attr():
 
 def test_canonical_path_class_no_module():
     """_canonical_path returns None when class has no __module__."""
-    from great_docs._renderer.introspect import _canonical_path
+    from great_docs._apiref.introspect import _canonical_path
 
     cls = type("DynClass", (), {})
     cls.__module__ = None
@@ -28790,13 +28790,13 @@ def test_dynamic_alias_nonexistent_attr_raises():
 
 def test_dynamic_alias_with_loader():
     """dynamic_alias accepts a shared loader."""
-    from great_docs._renderer._griffe import (
+    from great_docs._apiref._griffe import (
         GriffeLoader,
         LinesCollection,
         ModulesCollection,
         Parser,
     )
-    from great_docs._renderer.introspect import dynamic_alias
+    from great_docs._apiref.introspect import dynamic_alias
 
     loader = GriffeLoader(
         docstring_parser=Parser("numpy"),
@@ -30413,7 +30413,7 @@ def test_ast_transform_all():
 
 def test_ast_fields_example_code():
     """fields() returns dataclass field names for ExampleCode."""
-    from great_docs._renderer._ast import ExampleCode, fields
+    from great_docs._apiref._ast import ExampleCode, fields
 
     ec = ExampleCode("print(1)")
     result = fields(ec)
@@ -30422,7 +30422,7 @@ def test_ast_fields_example_code():
 
 def test_ast_fields_example_text():
     """fields() returns dataclass field names for ExampleText."""
-    from great_docs._renderer._ast import ExampleText, fields
+    from great_docs._apiref._ast import ExampleText, fields
 
     et = ExampleText("description")
     result = fields(et)
@@ -30431,7 +30431,7 @@ def test_ast_fields_example_text():
 
 def test_ast_fields_function():
     """fields() returns expected fields for dc.Function."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     func = dc.Function(name="my_func", lineno=1)
     result = fields(func)
@@ -30441,7 +30441,7 @@ def test_ast_fields_function():
 
 def test_ast_fields_attribute():
     """fields() returns expected fields for dc.Attribute."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     attr = dc.Attribute(name="x", lineno=1)
     result = fields(attr)
@@ -30451,7 +30451,7 @@ def test_ast_fields_attribute():
 
 def test_ast_fields_docstring():
     """fields() returns expected fields for dc.Docstring."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     ds_obj = dc.Docstring("Hello", parser="numpy")
     result = fields(ds_obj)
@@ -30461,7 +30461,7 @@ def test_ast_fields_docstring():
 
 def test_ast_fields_parameter():
     """fields() returns expected fields for dc.Parameter."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     param = dc.Parameter(name="x")
     result = fields(param)
@@ -30471,7 +30471,7 @@ def test_ast_fields_parameter():
 
 def test_ast_fields_docstring_parameter():
     """fields() returns expected fields for ds.DocstringParameter."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     dp = ds.DocstringParameter(name="x", description="a number", annotation="int")
     result = fields(dp)
@@ -30481,7 +30481,7 @@ def test_ast_fields_docstring_parameter():
 
 def test_ast_fields_docstring_named_element():
     """fields() returns expected fields for ds.DocstringNamedElement."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     ne = ds.DocstringNamedElement(name="x", description="desc")
     result = fields(ne)
@@ -30491,7 +30491,7 @@ def test_ast_fields_docstring_named_element():
 
 def test_ast_fields_docstring_section():
     """fields() returns expected fields for ds.DocstringSection."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     sec = ds.DocstringSectionText("hello")
     result = fields(sec)
@@ -30501,7 +30501,7 @@ def test_ast_fields_docstring_section():
 
 def test_ast_fields_alias():
     """fields() follows alias target for dc.Alias."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     mod = dc.Module(name="testmod")
     func = dc.Function(name="f", lineno=1)
@@ -30518,7 +30518,7 @@ def test_ast_fields_alias_unresolvable():
 
     from griffe import ModulesCollection
 
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     mc = ModulesCollection()
     mod = dc.Module(name="testmod")
@@ -30536,7 +30536,7 @@ def test_ast_fields_alias_unresolvable():
 
 def test_ast_fields_object():
     """fields() returns discovered attributes for dc.Object (Module)."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     mod = dc.Module(name="mymod")
     result = fields(mod)
@@ -30548,7 +30548,7 @@ def test_ast_fields_object():
 
 def test_ast_fields_layout_base():
     """fields() returns non-default fields for a LayoutBase subclass."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     auto = Auto(name="my_func")
     result = fields(auto)
@@ -30557,7 +30557,7 @@ def test_ast_fields_layout_base():
 
 def test_ast_fields_dict():
     """fields() returns dict keys."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     result = fields({"a": 1, "b": 2})
     assert result == ["a", "b"]
@@ -30565,7 +30565,7 @@ def test_ast_fields_dict():
 
 def test_ast_fields_list():
     """fields() returns list indices."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     result = fields([10, 20, 30])
     assert result == [0, 1, 2]
@@ -30573,7 +30573,7 @@ def test_ast_fields_list():
 
 def test_ast_fields_parameters():
     """fields() returns indices for dc.Parameters."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     params = dc.Parameters(dc.Parameter(name="a"), dc.Parameter(name="b"))
     result = fields(params)
@@ -30583,7 +30583,7 @@ def test_ast_fields_parameters():
 
 def test_ast_fields_none_for_unknown():
     """fields() returns None for unrecognized types."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     assert fields(42) is None
     assert fields("hello") is None
@@ -30722,7 +30722,7 @@ def test_ast_formatter_format_nested():
 
 def test_ast_preview_print(capsys):
     """preview() prints the formatted tree."""
-    from great_docs._renderer._ast import ExampleCode, preview
+    from great_docs._apiref._ast import ExampleCode, preview
 
     preview(ExampleCode("x = 1"))
     captured = capsys.readouterr()
@@ -30731,7 +30731,7 @@ def test_ast_preview_print(capsys):
 
 def test_ast_preview_as_string():
     """preview(as_string=True) returns the tree as a string."""
-    from great_docs._renderer._ast import ExampleCode, preview
+    from great_docs._apiref._ast import ExampleCode, preview
 
     result = preview(ExampleCode("x = 1"), as_string=True)
     assert isinstance(result, str)
@@ -30740,7 +30740,7 @@ def test_ast_preview_as_string():
 
 def test_ast_preview_max_depth():
     """preview() respects max_depth parameter."""
-    from great_docs._renderer._ast import ExampleCode, preview
+    from great_docs._apiref._ast import ExampleCode, preview
 
     result = preview(ExampleCode("x = 1"), max_depth=0, as_string=True)
     assert "ExampleCode" in result
@@ -30749,7 +30749,7 @@ def test_ast_preview_max_depth():
 
 def test_ast_preview_compact():
     """preview() respects compact parameter."""
-    from great_docs._renderer._ast import ExampleCode, preview
+    from great_docs._apiref._ast import ExampleCode, preview
 
     result = preview(ExampleCode("x = 1"), compact=True, as_string=True)
     assert "ExampleCode" in result
@@ -30780,7 +30780,7 @@ def test_ast_docstring_section_warnings():
 
 def test_ast_fields_docstring_element():
     """fields() returns expected fields for ds.DocstringElement."""
-    from great_docs._renderer._ast import fields
+    from great_docs._apiref._ast import fields
 
     # DocstringElement is the base; DocstringReturn is a subclass not matching
     # DocstringNamedElement or DocstringParameter
@@ -33581,7 +33581,7 @@ def test_emph_str():
 
 def test_image_str():
     """Image.__str__ renders markdown image syntax."""
-    from great_docs._renderer.pandoc.inlines import Image
+    from great_docs._apiref.pandoc.inlines import Image
 
     img = Image(caption="Logo", src="img.png")
     assert str(img) == "![Logo](img.png)"
@@ -33771,7 +33771,7 @@ def test_is_field_init_false_no_field():
 
 def test_canonical_path_with_type():
     """_canonical_path returns module.qualname for a type."""
-    from great_docs._renderer._tools import _canonical_path
+    from great_docs._apiref._tools import _canonical_path
 
     result = _canonical_path(int)
     assert result == "int"  # builtins returns just qualname
@@ -33782,7 +33782,7 @@ def test_canonical_path_with_type():
 
 def test_canonical_path_with_class():
     """_canonical_path returns full path for non-builtin type."""
-    from great_docs._renderer._tools import _canonical_path
+    from great_docs._apiref._tools import _canonical_path
 
     result = _canonical_path(Attr)
     assert "Attr" in result
@@ -33791,7 +33791,7 @@ def test_canonical_path_with_class():
 
 def test_canonical_path_with_instance():
     """_canonical_path handles non-type by using __class__."""
-    from great_docs._renderer._tools import _canonical_path
+    from great_docs._apiref._tools import _canonical_path
 
     result = _canonical_path("hello")
     assert result == "str"  # builtins
@@ -33817,7 +33817,7 @@ def test_render_type_object_with_type():
     """render_type_object renders a python type."""
 
     # Use a type from the package itself that griffe can find
-    result = render_type_object("great_docs._renderer.pandoc.components.Attr")
+    result = render_type_object("great_docs._apiref.pandoc.components.Attr")
     assert isinstance(result, str)
 
 
@@ -35135,7 +35135,7 @@ def test_renderdoc_source_link_with_github_url():
 
     # Patch package_info directly so we don't depend on os.environ ordering
     with patch(
-        "great_docs._renderer._render.doc.package_info",
+        "great_docs._apiref._render.doc.package_info",
         side_effect=lambda key: {
             "GITHUB_REPO_URL": "https://github.com/test/repo",
             "GIT_REF": "main",
@@ -35618,9 +35618,9 @@ def test_type_sections_items_combines_all():
     mock_render_cls = MagicMock(return_value=mock_render)
 
     with (
-        patch("great_docs._renderer.typing_information.griffe_to_doc") as mock_g2d,
+        patch("great_docs._apiref.typing_information.griffe_to_doc") as mock_g2d,
         patch(
-            "great_docs._renderer.typing_information.get_render_type", return_value=mock_render_cls
+            "great_docs._apiref.typing_information.get_render_type", return_value=mock_render_cls
         ),
     ):
         mock_g2d.return_value = MagicMock()
@@ -35645,9 +35645,9 @@ def test_type_sections_post_init_protocols_no_summary():
     mock_render_cls = MagicMock(return_value=mock_render)
 
     with (
-        patch("great_docs._renderer.typing_information.griffe_to_doc"),
+        patch("great_docs._apiref.typing_information.griffe_to_doc"),
         patch(
-            "great_docs._renderer.typing_information.get_render_type", return_value=mock_render_cls
+            "great_docs._apiref.typing_information.get_render_type", return_value=mock_render_cls
         ),
     ):
         ts = TypeSections(
@@ -35668,9 +35668,9 @@ def test_type_sections_post_init_typevars_no_sig_name():
     mock_render_cls = MagicMock(return_value=mock_render)
 
     with (
-        patch("great_docs._renderer.typing_information.griffe_to_doc"),
+        patch("great_docs._apiref.typing_information.griffe_to_doc"),
         patch(
-            "great_docs._renderer.typing_information.get_render_type", return_value=mock_render_cls
+            "great_docs._apiref.typing_information.get_render_type", return_value=mock_render_cls
         ),
     ):
         ts = TypeSections(
@@ -35691,9 +35691,9 @@ def test_type_sections_post_init_typealiases_settings():
     mock_render_cls = MagicMock(return_value=mock_render)
 
     with (
-        patch("great_docs._renderer.typing_information.griffe_to_doc"),
+        patch("great_docs._apiref.typing_information.griffe_to_doc"),
         patch(
-            "great_docs._renderer.typing_information.get_render_type", return_value=mock_render_cls
+            "great_docs._apiref.typing_information.get_render_type", return_value=mock_render_cls
         ),
     ):
         ts = TypeSections(
@@ -35716,9 +35716,9 @@ def test_type_sections_render_body_protocols_section():
     mock_render_cls = MagicMock(return_value=mock_render)
 
     with (
-        patch("great_docs._renderer.typing_information.griffe_to_doc"),
+        patch("great_docs._apiref.typing_information.griffe_to_doc"),
         patch(
-            "great_docs._renderer.typing_information.get_render_type", return_value=mock_render_cls
+            "great_docs._apiref.typing_information.get_render_type", return_value=mock_render_cls
         ),
     ):
         ts = TypeSections(
@@ -35744,9 +35744,9 @@ def test_type_sections_render_body_typevars_section():
     mock_render_cls = MagicMock(return_value=mock_render)
 
     with (
-        patch("great_docs._renderer.typing_information.griffe_to_doc"),
+        patch("great_docs._apiref.typing_information.griffe_to_doc"),
         patch(
-            "great_docs._renderer.typing_information.get_render_type", return_value=mock_render_cls
+            "great_docs._apiref.typing_information.get_render_type", return_value=mock_render_cls
         ),
     ):
         ts = TypeSections(
@@ -35770,9 +35770,9 @@ def test_type_sections_render_body_typealiases_section():
     mock_render_cls = MagicMock(return_value=mock_render)
 
     with (
-        patch("great_docs._renderer.typing_information.griffe_to_doc"),
+        patch("great_docs._apiref.typing_information.griffe_to_doc"),
         patch(
-            "great_docs._renderer.typing_information.get_render_type", return_value=mock_render_cls
+            "great_docs._apiref.typing_information.get_render_type", return_value=mock_render_cls
         ),
     ):
         ts = TypeSections(
@@ -35798,9 +35798,9 @@ def test_type_sections_render_body_all_sections():
     mock_render_cls = MagicMock(return_value=mock_render)
 
     with (
-        patch("great_docs._renderer.typing_information.griffe_to_doc"),
+        patch("great_docs._apiref.typing_information.griffe_to_doc"),
         patch(
-            "great_docs._renderer.typing_information.get_render_type", return_value=mock_render_cls
+            "great_docs._apiref.typing_information.get_render_type", return_value=mock_render_cls
         ),
     ):
         ts = TypeSections(
@@ -35872,22 +35872,22 @@ def test_type_information_sections_calls_get_object():
     mock_render_cls = MagicMock(return_value=mock_render_obj)
 
     with (
-        patch("great_docs._renderer.typing_information.get_object", return_value=mock_module),
+        patch("great_docs._apiref.typing_information.get_object", return_value=mock_module),
         patch(
-            "great_docs._renderer.typing_information.is_protocol",
+            "great_docs._apiref.typing_information.is_protocol",
             side_effect=lambda m: m is mock_proto,
         ),
         patch(
-            "great_docs._renderer.typing_information.is_typevar",
+            "great_docs._apiref.typing_information.is_typevar",
             side_effect=lambda m: m is mock_tv,
         ),
         patch(
-            "great_docs._renderer.typing_information.is_typealias",
+            "great_docs._apiref.typing_information.is_typealias",
             side_effect=lambda m: m is mock_alias,
         ),
-        patch("great_docs._renderer.typing_information.griffe_to_doc"),
+        patch("great_docs._apiref.typing_information.griffe_to_doc"),
         patch(
-            "great_docs._renderer.typing_information.get_render_type", return_value=mock_render_cls
+            "great_docs._apiref.typing_information.get_render_type", return_value=mock_render_cls
         ),
     ):
         ti = TypeInformation(module_path="mypkg.types", api_ref=mock_builder)
@@ -35918,13 +35918,13 @@ def test_type_information_sections_item_uris():
     mock_render_cls = MagicMock(return_value=mock_render_obj)
 
     with (
-        patch("great_docs._renderer.typing_information.get_object", return_value=mock_module),
-        patch("great_docs._renderer.typing_information.is_protocol", return_value=False),
-        patch("great_docs._renderer.typing_information.is_typevar", return_value=False),
-        patch("great_docs._renderer.typing_information.is_typealias", return_value=True),
-        patch("great_docs._renderer.typing_information.griffe_to_doc"),
+        patch("great_docs._apiref.typing_information.get_object", return_value=mock_module),
+        patch("great_docs._apiref.typing_information.is_protocol", return_value=False),
+        patch("great_docs._apiref.typing_information.is_typevar", return_value=False),
+        patch("great_docs._apiref.typing_information.is_typealias", return_value=True),
+        patch("great_docs._apiref.typing_information.griffe_to_doc"),
         patch(
-            "great_docs._renderer.typing_information.get_render_type", return_value=mock_render_cls
+            "great_docs._apiref.typing_information.get_render_type", return_value=mock_render_cls
         ),
     ):
         ti = TypeInformation(module_path="mypkg.sub", api_ref=mock_builder)
@@ -35945,7 +35945,7 @@ def test_type_information_content_has_meta_and_sections():
     mock_module = MagicMock()
     mock_module.members = {}
 
-    with patch("great_docs._renderer.typing_information.get_object", return_value=mock_module):
+    with patch("great_docs._apiref.typing_information.get_object", return_value=mock_module):
         ti = TypeInformation(module_path="mypkg.types", api_ref=mock_builder)
         content = ti.content
 
@@ -35963,7 +35963,7 @@ def test_type_information_str_delegates_to_content():
     mock_module = MagicMock()
     mock_module.members = {}
 
-    with patch("great_docs._renderer.typing_information.get_object", return_value=mock_module):
+    with patch("great_docs._apiref.typing_information.get_object", return_value=mock_module):
         ti = TypeInformation(module_path="mypkg.types", api_ref=mock_builder)
         result = str(ti)
 
@@ -35982,7 +35982,7 @@ def test_type_information_write_creates_file():
         mock_module = MagicMock()
         mock_module.members = {}
 
-        with patch("great_docs._renderer.typing_information.get_object", return_value=mock_module):
+        with patch("great_docs._apiref.typing_information.get_object", return_value=mock_module):
             ti = TypeInformation(module_path="mypkg.types", api_ref=mock_builder)
 
             # Ensure the output directory exists
@@ -36016,13 +36016,13 @@ def test_type_information_write_extends_builder_items():
         mock_render_cls = MagicMock(return_value=mock_render_obj)
 
         with (
-            patch("great_docs._renderer.typing_information.get_object", return_value=mock_module),
-            patch("great_docs._renderer.typing_information.is_protocol", return_value=False),
-            patch("great_docs._renderer.typing_information.is_typevar", return_value=True),
-            patch("great_docs._renderer.typing_information.is_typealias", return_value=False),
-            patch("great_docs._renderer.typing_information.griffe_to_doc"),
+            patch("great_docs._apiref.typing_information.get_object", return_value=mock_module),
+            patch("great_docs._apiref.typing_information.is_protocol", return_value=False),
+            patch("great_docs._apiref.typing_information.is_typevar", return_value=True),
+            patch("great_docs._apiref.typing_information.is_typealias", return_value=False),
+            patch("great_docs._apiref.typing_information.griffe_to_doc"),
             patch(
-                "great_docs._renderer.typing_information.get_render_type",
+                "great_docs._apiref.typing_information.get_render_type",
                 return_value=mock_render_cls,
             ),
         ):
